split mitmproxy.flow into mitmproxy.flow.*
